+++ libvast/src/system/spawn_arguments.cpp
@@ -31,15 +31,21 @@
 
 namespace vast::system {
 
-caf::expected<expression> normalized_and_valided(const spawn_arguments& args) {
-  if (args.empty())
+caf::expected<expression>
+normalized_and_validated(std::vector<std::string>::const_iterator begin,
+                         std::vector<std::string>::const_iterator end) {
+  if (begin == end)
     return make_error(ec::syntax_error, "no query expression given");
-  if (auto e = to<expression>(caf::join(args.invocation.arguments.begin(),
-                                        args.invocation.arguments.end(), " "));
-      !e)
-    return std::move(e.error());
-  else
+  if (auto e = to<expression>(caf::join(begin, end, " ")))
     return normalize_and_validate(*e);
+  else
+    return std::move(e.error());
+}
+
+caf::expected<expression>
+normalized_and_validated(const spawn_arguments& args) {
+  auto& arguments = args.invocation.arguments;
+  return normalized_and_validated(arguments.begin(), arguments.end());
 }
 
 caf::expected<caf::optional<schema>> read_schema(const spawn_arguments& args) {