+++ libvast/vast/system/pivot_command.hpp
@@ -15,18 +15,10 @@
 
 #include "vast/command.hpp"
 
-#include <caf/fwd.hpp>
-
 namespace vast::system {
 
-/// Reads query from input file, STDIN or CLI arguments.
-/// @param invocation The command line in parsed form.
-/// @param file_option The option name that is used to pass the query by file
-///                    instead of as command line argument(s).
-/// @param argument_offset The number of argumetns to skip before the query.
-/// @returns The query string or an error.
-caf::expected<std::string>
-read_query(const command::invocation& invocation, std::string_view file_option,
-           size_t argument_offset = 0);
+/// Implementation for pivot command
+caf::message
+pivot_command(const command::invocation& invocation, caf::actor_system& sys);
 
 } // namespace vast::system